requirements.txt
pillow
keyboard
pystray
pywin32